created separate node file for launching, refactoring
