+++ velocity_topic_watchdog/CMakeLists.txt
@@ -133,7 +133,9 @@ include_directories(
 ## Declare a C++ executable
 ## With catkin_make all packages are built within a single CMake context
 ## The recommended prefix ensures that target names across packages don't collide
-add_executable(velocity_topic_watchdog src/velocity_topic_watchdog.cpp include/velocity_topic_watchdog/velocity_topic_watchdog.h)
+add_executable(velocity_topic_watchdog src/VelocityTopicWatchdog.cpp
+        include/velocity_topic_watchdog/VelocityTopicWatchdog.hpp
+        src/velocity_watchdog_node.cpp)
 
 ## Rename C++ executable without prefix
 ## The above recommended prefix causes long target names, the following renames the
